Move apps, admin and missing views
